+++ src/viewContext.ts
@@ -1,8 +1,8 @@
 import type { TSettingsFilter } from "./filtration";
-import type { EWidgetFilterMode } from "./settings/values";
+import type { TWidgetFilterMode } from "./settings/values";
 
 /** Контекст с данными образа (будет заполняться по мере необходимости) */
 export interface IViewContext {
   filters: TSettingsFilter[];
-  filterMode: EWidgetFilterMode;
+  filterMode: TWidgetFilterMode;
 }